requirements.txt
streamlit>=1.29.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.12.0
PyPDF2>=3.0.1
python-dotenv>=1.0.0
langchain-community>=0.0.20
reportlab>=4.0.0
